fixed a few tests
