+++ src/test/java/com/revature/AuthControllerTest.java
@@ -1,14 +1,16 @@
-package com.revature.controllers;
+package com.revature;
 
 import com.revature.annotations.AuthRestriction;
+import com.revature.controllers.AuthController;
 import com.revature.dtos.Jwt;
 import com.revature.dtos.LoginRequest;
 import com.revature.dtos.PasswordChange;
 import com.revature.dtos.RegisterRequest;
+import com.revature.models.Question;
 import com.revature.models.User;
 import com.revature.services.AuthService;
+import com.revature.services.QuestionService;
 import com.revature.services.UserService;
-import net.bytebuddy.dynamic.DynamicType;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -18,6 +20,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Optional;
 
 import static org.mockito.Mockito.when;
@@ -34,6 +38,9 @@ public class AuthControllerTest {
     @Mock
     UserService userServiceMock;
 
+    @Mock
+    QuestionService questionServiceMock;
+
     @Test
     void login_test(){
         Jwt jwt = new Jwt("jwt", "access");
@@ -48,9 +55,10 @@ public class AuthControllerTest {
     @Test
     void register_test(){
         User mockUser = new User(0, "email", "pass", "first", "last", "img", AuthRestriction.USER);
-        RegisterRequest rr = new RegisterRequest("email", "pass", "first", "last", "img");
+        Question q = new Question();
+        RegisterRequest rr = new RegisterRequest(mockUser, q);
 
-        when(authServiceMock.register(mockUser)).thenReturn(mockUser);
+        when(authServiceMock.register(mockUser, q)).thenReturn(mockUser);
 
         ResponseEntity<User> re = authController.register(rr);
 
@@ -61,19 +69,52 @@ public class AuthControllerTest {
     void change_credential_works_test(){
         PasswordChange ps = new PasswordChange("user", "newPass");
         User testUser = new User();
+        testUser.setId(1);
+        Question testQuestion = new Question();
+        testQuestion.setAnswer("answer");
         when(userServiceMock.findByUsername("user")).thenReturn(Optional.of(testUser));
+        when(questionServiceMock.findByUserid(1)).thenReturn(Optional.of(testQuestion));
 
-        ResponseEntity<User> re = authController.changeCredential(ps);
+        Map<String,String> m = new HashMap<>();
+        m.putIfAbsent("username", "user");
+        m.putIfAbsent("password", "newPass");
+        m.putIfAbsent("answer", "answer");
+        ResponseEntity<User> re = authController.changeCredential(m);
         Assertions.assertEquals(HttpStatus.CREATED, re.getStatusCode());
     }
 
     @Test
     void change_credential_fails_test(){
         PasswordChange ps = new PasswordChange("user", "newPass");
+        User testUser = new User();
+        testUser.setId(1);
+        Question testQuestion = new Question();
+        testQuestion.setAnswer("badAnswer");
+        when(userServiceMock.findByUsername("user")).thenReturn(Optional.of(testUser));
+        when(questionServiceMock.findByUserid(1)).thenReturn(Optional.of(testQuestion));
+
+        Map<String,String> m = new HashMap<>();
+        m.putIfAbsent("username", "user");
+        m.putIfAbsent("password", "newPass");
+        m.putIfAbsent("answer", "answer");
+        ResponseEntity<User> re = authController.changeCredential(m);
+        Assertions.assertEquals(HttpStatus.BAD_REQUEST, re.getStatusCode());
+    }
+    @Test
+    void change_credential_not_found_test(){
+        PasswordChange ps = new PasswordChange("user", "newPass");
+        User testUser = new User();
+        testUser.setId(1);
+        Question testQuestion = new Question();
+        testQuestion.setAnswer("answer");
         when(userServiceMock.findByUsername("user")).thenReturn(Optional.empty());
 
-        ResponseEntity<User> re = authController.changeCredential(ps);
-        Assertions.assertEquals(HttpStatus.BAD_REQUEST, re.getStatusCode());
+        Map<String,String> m = new HashMap<>();
+        m.putIfAbsent("username", "user");
+        m.putIfAbsent("password", "newPass");
+        m.putIfAbsent("answer", "answer");
+        ResponseEntity<User> re = authController.changeCredential(m);
+        Assertions.assertEquals(HttpStatus.NOT_FOUND, re.getStatusCode());
     }
 
     @Test
