+++ src/test/java/com/revature/AuthServiceTests.java
@@ -1,4 +1,4 @@
-package com.revature.services;
+package com.revature;
 
 import com.revature.annotations.AuthRestriction;
 import com.revature.dtos.LoginRequest;
@@ -7,6 +7,7 @@ import com.revature.exceptions.UserExistsException;
 import com.revature.exceptions.UserNotFoundException;
 import com.revature.models.Question;
 import com.revature.models.User;
+import com.revature.services.AuthService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
