+++ src/test/java/com/revature/NotAuthorizedTests.java
@@ -1,36 +1,35 @@
-package com.revature.except;
+package com.revature;
 
-import com.revature.exceptions.NotLoggedInException;
+import com.revature.exceptions.NotAuthorizedException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-public class NotLoggedInTests {
+public class NotAuthorizedTests {
 
     @Test
     void message_test(){
-        NotLoggedInException e = new NotLoggedInException();
-        NotLoggedInException ex = new NotLoggedInException("message");
+        NotAuthorizedException ex = new NotAuthorizedException("message");
         Assertions.assertEquals("message", ex.getMessage());
     }
     @Test
     void cause_test(){
         RuntimeException e = new RuntimeException("cause");
-        NotLoggedInException ex = new NotLoggedInException(e);
+        NotAuthorizedException ex = new NotAuthorizedException(e);
         Assertions.assertEquals(e, ex.getCause());
     }
     @Test
     void cause_and_message_test(){
         RuntimeException e = new RuntimeException("cause");
-        NotLoggedInException ex = new NotLoggedInException("message",e);
+        NotAuthorizedException ex = new NotAuthorizedException("message",e);
         Assertions.assertEquals("message", ex.getMessage());
         Assertions.assertEquals(e, ex.getCause());
     }
     @Test
     void all_params_test(){
         RuntimeException e = new RuntimeException("cause");
-        NotLoggedInException ex = new NotLoggedInException("message",e,true,true);
+        NotAuthorizedException ex = new NotAuthorizedException("message",e,true,true);
         Assertions.assertEquals("message", ex.getMessage());
         Assertions.assertEquals(e, ex.getCause());
     }
