+++ src/test/java/com/revature/PasswordMismatchTests.java
@@ -1,35 +1,35 @@
-package com.revature.except;
+package com.revature;
 
-import com.revature.exceptions.NotAuthorizedException;
+import com.revature.exceptions.PasswordMismatchException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-public class NotAuthorizedTests {
+public class PasswordMismatchTests {
 
     @Test
     void message_test(){
-        NotAuthorizedException ex = new NotAuthorizedException("message");
+        PasswordMismatchException ex = new PasswordMismatchException("message");
         Assertions.assertEquals("message", ex.getMessage());
     }
     @Test
     void cause_test(){
         RuntimeException e = new RuntimeException("cause");
-        NotAuthorizedException ex = new NotAuthorizedException(e);
+        PasswordMismatchException ex = new PasswordMismatchException(e);
         Assertions.assertEquals(e, ex.getCause());
     }
     @Test
     void cause_and_message_test(){
         RuntimeException e = new RuntimeException("cause");
-        NotAuthorizedException ex = new NotAuthorizedException("message",e);
+        PasswordMismatchException ex = new PasswordMismatchException("message",e);
         Assertions.assertEquals("message", ex.getMessage());
         Assertions.assertEquals(e, ex.getCause());
     }
     @Test
     void all_params_test(){
         RuntimeException e = new RuntimeException("cause");
-        NotAuthorizedException ex = new NotAuthorizedException("message",e,true,true);
+        PasswordMismatchException ex = new PasswordMismatchException("message",e,true,true);
         Assertions.assertEquals("message", ex.getMessage());
         Assertions.assertEquals(e, ex.getCause());
     }
