+++ src/test/java/com/revature/ProfileControllerTests.java
@@ -1,5 +1,6 @@
-package com.revature.controllers;
+package com.revature;
 
+import com.revature.controllers.ProfileController;
 import com.revature.dtos.ProfileInfo;
 import com.revature.exceptions.NotAuthorizedException;
 import com.revature.exceptions.UserNotFoundException;
